refactor: app structure at load time
