+++ src/gridGL/QuadraticGrid.tsx
@@ -1,5 +1,4 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { useLoading } from '../contexts/LoadingContext';
 import { gridInteractionStateAtom } from '../atoms/gridInteractionStateAtom';
 import { editorInteractionStateAtom } from '../atoms/editorInteractionStateAtom';
 import { useRecoilState } from 'recoil';
@@ -21,8 +20,6 @@ let mouseIsDown = false;
 let spaceIsDown = false;
 
 export default function QuadraticGrid(props: IProps) {
-  const { loading } = useLoading();
-
   const [container, setContainer] = useState<HTMLDivElement>();
   const containerRef = useCallback((node: HTMLDivElement | null) => {
     if (node) setContainer(node);
@@ -106,8 +103,6 @@ export default function QuadraticGrid(props: IProps) {
     setEditorInteractionState,
     app: props.app,
   });
-
-  if (loading) return null;
 
   return (
     <div
