+++ src/quadratic/QuadraticApp.tsx
@@ -1,21 +1,36 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import QuadraticUIContext from '../ui/QuadraticUIContext';
-import { RecoilRoot } from 'recoil';
-import { useLoading } from '../contexts/LoadingContext';
 import { QuadraticLoading } from '../ui/loading/QuadraticLoading';
 import { loadPython } from '../grid/computations/python/loadPython';
 import { AnalyticsProvider } from './AnalyticsProvider';
 import { loadAssets } from '../gridGL/loadAssets';
 import { IS_READONLY_MODE } from '../constants/app';
 import { debugSkipPythonLoad } from '../debugFlags';
-import { localFiles } from '../grid/sheet/localFiles';
 import init, { hello } from 'quadratic-core';
 import { useGridSettings } from '../ui/menus/TopBar/SubMenus/useGridSettings';
+import { SheetController } from '../grid/controller/sheetController';
+import { useLocalFiles } from '../storage/useLocalFiles';
+import { PixiApp } from '../gridGL/pixiApp/PixiApp';
+
+type loadItem = 'pixi' | 'local-files' | 'wasm-rust' | 'wasm-python';
+const THINGS_TO_LOAD = ['pixi', 'local-files', 'wasm-rust', 'wasm-python'] as loadItem[];
 
 export const QuadraticApp = () => {
-  const { loading, incrementLoadingCount } = useLoading();
+  const [loading, setLoading] = useState(true);
+  const [thingsLoaded, setThingsLoaded] = useState<loadItem[]>([]);
+  const didMount = useRef(false);
   const { setPresentationMode } = useGridSettings();
   const [settingsReset, setSettingsReset] = useState(false);
+  const [sheetController] = useState<SheetController>(new SheetController());
+  const localFiles = useLocalFiles(sheetController);
+  const [app] = useState(() => new PixiApp(sheetController, localFiles.save));
+  const { initialize } = localFiles;
+
+  useEffect(() => {
+    if (THINGS_TO_LOAD.every((item) => thingsLoaded.includes(item))) {
+      setLoading(false);
+    }
+  }, [thingsLoaded]);
 
   // reset presentation mode when app starts
   useEffect(() => {
@@ -27,36 +42,38 @@ export const QuadraticApp = () => {
 
   // Loading Effect
   useEffect(() => {
-    if (loading) {
-      if (!IS_READONLY_MODE && !debugSkipPythonLoad) {
-        loadPython().then(() => {
-          incrementLoadingCount();
-        });
-      } else {
-        incrementLoadingCount();
-      }
-      loadAssets().then(() => {
-        incrementLoadingCount();
+    // Ensure this only runs once
+    if (didMount.current) return;
+    didMount.current = true;
+
+    if (!IS_READONLY_MODE && !debugSkipPythonLoad) {
+      loadPython().then(() => {
+        setThingsLoaded((old) => ['wasm-python', ...old]);
       });
-      // load Rust wasm
-      init().then(() => {
-        hello(); // let Rust say hello to console
-        incrementLoadingCount();
-      });
-      localFiles.initialize().then(() => {
-        incrementLoadingCount();
-      });
+    } else {
+      setThingsLoaded((old) => ['wasm-python', ...old]);
     }
-  }, [loading, incrementLoadingCount]);
+    loadAssets().then(() => {
+      setThingsLoaded((old) => ['pixi', ...old]);
+    });
+    // load Rust wasm
+    init().then(() => {
+      hello(); // let Rust say hello to console
+      setThingsLoaded((old) => ['wasm-rust', ...old]);
+    });
+    initialize().then(() => {
+      setThingsLoaded((old) => ['local-files', ...old]);
+    });
+  }, [initialize]);
 
   return (
-    <RecoilRoot>
+    <>
       {/* Provider for Analytics. Only used when running in Quadratic Cloud. */}
       <AnalyticsProvider></AnalyticsProvider>
       {/* Provider of All React UI Components */}
-      {!loading && <QuadraticUIContext />}
+      {!loading && <QuadraticUIContext {...{ sheetController, localFiles, app }} />}
       {/* Loading screen */}
       {loading && <QuadraticLoading></QuadraticLoading>}
-    </RecoilRoot>
+    </>
   );
 };
