+++ src/ui/QuadraticUIContext.tsx
@@ -1,7 +1,4 @@
-import { createContext, useState } from 'react';
-import { SheetController } from '../grid/controller/sheetController';
-import { PixiApp } from '../gridGL/pixiApp/PixiApp';
-import { useLocalFiles } from '../storage/useLocalFiles';
+import { createContext } from 'react';
 import { LocalFiles } from '../storage/useLocalFiles';
 import QuadraticUI from './QuadraticUI';
 
@@ -11,11 +8,7 @@ export const LocalFilesContext = createContext<LocalFiles>({} as LocalFiles);
 // export const PixiAppContext = createContext<PixiApp>({} as PixiApp);
 // export const SheetControllerContext = createContext<SheetController>({} as SheetController);
 
-export default function QuadraticUIContext() {
-  const [sheetController] = useState<SheetController>(new SheetController());
-  const localFiles = useLocalFiles(sheetController);
-  const [app] = useState(() => new PixiApp(sheetController, localFiles.save));
-
+export default function QuadraticUIContext({ sheetController, localFiles, app }: any) {
   return (
     <LocalFilesContext.Provider value={localFiles}>
       {/* <PixiAppContext.Provider value={app}>
